fix: prevent Local 502 error by using unoptimized images
Local 6.6+ uses Electron 21, which implements a new memory cage feature:

https://www.electronjs.org/blog/v8-memory-cage

The memory cage prevents Node.js modules such as sharp from working:

https://github.com/lovell/sharp/issues/3384

Sharp is used by Next.js to optimize images.

The result is that Local produces a 502 when users create new sites
using this blueprint, due to the optimized logo image in the header and
in featured images.

Using the `unoptimized` prop prevents this.

The fix could be removed when Electron updates with the patched version
of Node.js that allows sharp's workaround to function, and when Local
has updated to that Electron version.